Working tunestarter generation
