+++ tunestarter/__init__.py
@@ -1,5 +1,3 @@
-from .tune import *
-from .set import *
 from .tunestarter import *
 
 __all__ = [
